+++ package.json
@@ -44,7 +44,7 @@
   "private": true,
   "scripts": {
     "build": "next build",
-    "dev": "next dev -p 3009",
+    "dev": "next dev -p 3010",
     "lint": "next lint",
     "start": "next start"
   },
